+++ src/FunderMaps.Core/Interfaces/Repositories/ITrackingRepository.cs
@@ -12,14 +12,16 @@ namespace FunderMaps.Core.Interfaces.Repositories
         /// <summary>
         ///     Process analysis product usage.
         /// </summary>
-        /// <param name="userId">Internal user id.</param>
+        /// <param name="userId">User identifier.</param>
+        /// <param name="product">Product type.</param>
         /// <param name="itemCount">Result set item count.</param>
         Task ProcessAnalysisUsageAsync(Guid userId, AnalysisProductType product, uint itemCount = 1);
 
         /// <summary>
         ///     Process statistics product usage.
         /// </summary>
-        /// <param name="userId">Internal user id.</param>
+        /// <param name="userId">User identifier.</param>
+        /// <param name="product">Product type.</param>
         /// <param name="itemCount">Result set item count.</param>
         Task ProcessStatisticsUsageAsync(Guid userId, StatisticsProductType product, uint itemCount = 1);
     }
